db/requirements.txt
fastapi==0.109.0
uvicorn==0.22.0
clickhouse-connect==0.7.7
kafka-python==2.0.2